# Read many values from the cache in a single call