--- a/app/views/articles/_full_comment_area.html.erb
+++ b/app/views/articles/_full_comment_area.html.erb
@@ -1,36 +1,40 @@
-<% cache("whole-comment-area-#{@article.id}-#{@article.last_comment_at}-#{@article.show_comments}", expires_in: 2.hours) do %>
-  <section id="comments" data-updated-at="<%= Time.current %>" class="text-padding mb-4 border-t-1 border-0 border-solid border-base-10">
-    <% if @article.show_comments %>
-      <header class="relative flex justify-between items-center mb-6">
-        <h2 class="crayons-subtitle-1">Discussion <span class="js-comments-count" data-comments-count="<%= @article.comments_count %>">(<%= @article.comments_count %>)</span></h2>
-        <div id="comment-subscription">
-          <div role="presentation" class="crayons-btn-group">
-            <span class="crayons-btn crayons-btn--outlined">Subscribe</span>
-          </div>
-        </div>
-      </header>
-
-      <div
-        id="comments-container"
-        data-testid="comments-container"
-        data-commentable-id="<%= @article.id %>"
-        data-commentable-type="Article"
-        data-has-recent-comment-activity="<%= @article.has_recent_comment_activity? %>">
-        <%= render "/comments/form", commentable: @article, commentable_type: "Article" %>
-
-        <div class="comments" id="comment-trees-container">
-          <% if @article.comments_count > 0 %>
-            <% Comment.tree_for(@article, @comments_to_show_count).each do |comment, sub_comments| %>
-              <% cache ["comment_root_cached_tree", comment] do %>
-                <%= tree_for(comment, sub_comments, @article) %>
-              <% end %>
-            <% end %>
-          <% end %>
+<section id="comments" data-updated-at="<%= Time.current %>" class="text-padding mb-4 border-t-1 border-0 border-solid border-base-10">
+  <% if @article.show_comments %>
+    <header class="relative flex justify-between items-center mb-6">
+      <h2 class="crayons-subtitle-1">
+        Discussion
+        <span class="js-comments-count" data-comments-count="<%= @article.comments_count %>">
+          (<%= @article.comments_count %>)
+        </span>
+      </h2>
+      <div id="comment-subscription">
+        <div role="presentation" class="crayons-btn-group">
+          <span class="crayons-btn crayons-btn--outlined">Subscribe</span>
         </div>
       </div>
+    </header>
 
-      <%= render "articles/comments_actions" %>
-    <% end %>
-  </section>
-<% end %>
+    <div
+      id="comments-container"
+      data-testid="comments-container"
+      data-commentable-id="<%= @article.id %>"
+      data-commentable-type="Article"
+      data-has-recent-comment-activity="<%= @article.has_recent_comment_activity? %>">
+      <%= render "/comments/form", commentable: @article, commentable_type: "Article" %>
+
+      <div class="comments" id="comment-trees-container">
+        <% if @article.comments_count > 0 %>
+          <% Comment.tree_for(@article, @comments_to_show_count).tap do |comments| %>
+            <% keys = comments.keys.map { |comment| ["comment_root_cached_tree", comment] } %>
+            <% cache_multi(keys, expires_in: 2.hours) do |_prefix, comment| %>
+              <%= tree_for(comment, comments.fetch(comment), @article) %>
+            <% end %>
+          <% end %>
+        <% end %>
+      </div>
+    </div>
+
+    <%= render "articles/comments_actions" %>
+  <% end %>
+</section>
 <% return if @warm_only %>