@citizen428 2021-05-11
> Doesn't look like render cached: true worked. 

Adding `cached: true` by itself may not work, I think you also need to add a `collection:` argument. That may require some more changes to the partial though so maybe is not worth the effort, I'm not sure.