@jgaskins 2021-05-11
@citizen428 [This line](https://github.com/forem/forem/compare/main...optimize-comment-rendering#diff-109f0ab13eba9c122830f97156302a597f24322c84f37d2ac5667a130d4f69eaR67) passes all of the top-level comments in as a collection. AFAICT `render things` and `render collection: things` should be equivalent — the main difference seems to be whether you need to pass an explicit `locals` vs the implicit form that this method uses, but honestly I don't remember the details. `render` is wildly overloaded in Rails helpers and it makes me wish we had _actual_ method overloading in Ruby. :joy: