@citizen428 2021-05-12
@jgaskins Agree, `render` is a hot mess and it's kind of hard to keep it all in one's head especially when you add things like collection caching, etc.